requirements.txt
pyyaml
pyqt5  # for dashboard UI
requests  # for LLM API integration 